Fix subdirectory header error, add input handling

+++ src/visual/gameObject.h
@@ -4,15 +4,15 @@
 
 class GameObject {
 public:
-    GameObject(Sprite* sprite, int x, int y, int width, int height);
+    GameObject(Sprite* sprite, float x, float y, int width, int height);
     ~GameObject();
     void update(float dt);
     void render(Renderer* renderer);
 
 private:
     Sprite* sprite;
-    int x;
-    int y;
+    float x;
+    float y;
     int width;
     int height;
 };
